Add label_scores and original_text properties to the Post class
Add label_scores and original_text properties to the Post class.
These properties are somewhat expensive to compute, so they're
computed lazily and cached when first accessed. With the new
properties on the Post object, the logic to compute the labels can
be removed from the make-dataset.py script. Similarly, we can use
the original text in the dataset rather than the selftext which is
occasionally edited in response to people's comments and labels.
Using the original text avoids leaking information about the
comments / label into the dataset instances.

Also, add in an additional post filter to require that the post
can successfully find it's original text. Looking at examples where
the original text cannot be found, they are mostly deleted posts,
cross-posts (with no selftext), and meta posts. So, filtering on
being able to find the original text improves the quality of the
resulting data.