Fix bug where analyze-performance.py would fail on joins
analyze-performance.py would attempt to join every key in a labeled
subset with the dataset. If a key didn't exist (for example the key
is for an instance that has been filtered out in the current
dataset version) then the script would crash with a KeyError. As
long as labeled subsets were sampled randomly, we can safely ignore
keys that fail to join in the current version and still obtain
unbiased estimates.