Refactor bin/analyze-performance.py and add make_confusion_matrix_str
Refactor analyze-performance.py to fill in a pre-defined report
template defined at the top of the script rather than print the
report piece by piece. This implementation style makes it easy to
edit the template and introduce new sections, as well as see how
the overall template will look.

Similarly, factor out the confusion matrix pretty printing logic
into it's own function so it can be reused in different scripts
and analyses in the future.