Move filtering logic onto the data models
Remove the filters as a separate module and place all the filtering
logic on the data modeling classes. This refactor makes the filtering
attributes easily accessible for other purposes and keeps the code
operating on the data with the definitions of that data.
Additionally, some new filters have been introduced.