Add the socialnorms.dataset subpackage
Add the socialnorms.dataset subpackage to provide code for working
with socialnorms in PyTorch. This code includes a PyTorch dataset
reader for socialnorms and a transform that can convert a pair of
text features to the proper format for the BERT model, as well as
a transform for composing other transforms. These transforms may
be used in combination with the dataset reader to handle all the
pre-processing necessary to use BERT.