Work around a bug in spaCy and add tests for StyleFeaturizer
spaCy has a bug where running a pipe with empty documents in some
of the batches will cause an assertion error to be thrown. Work
around the bug by replacing empty documents with a special token
(with a randomly generated component to avoid collisions), and
detecting this token later to invoke the correct behavior. Also,
refactor the testing code for StyleFeaturizer to make it easy to
add a regression test for this bug.