Add readers for the socialnorms benchmark and refactor tests
Add two readers for the socialnorms benchmark dataset along with
tests for the new readers, and refactor the old tests in order to
accommodate this new addition. In particular, move the socialnorms
corpus test fixture (tests/fixtures/socialnorms-easy) to a new path
(tests/fixtures/corpus-easy) and add a new fixture providing an
easy version of the benchmark.

Similarly, factor out the logic for copying package resources to
disk and modify the baseline test utilities to support testing
shallow baseline models on either dataset (the benchmark or the
corpus).