Factor out the code for fine-tuning LMs into its own function
Factor out the code for fine-tuning LMs into its own function since
we'll need that logic as a function in order to perform
hyper-parameter tuning. Moving that training code into the
baselines package also means that the configuration for the
benchmark and corpus baselines has to move out of
baselines/__init__.py and into new files, since the baselines.train
module cannot import the baselines module.