Improve label extraction from comments
Improve label extraction from comments by modifying the patterns
and adding conflict resolution logic for when multiple labels are
extracted. In cases where multiple labels are extracted, usually
the first label is the correct one unless there is some lexical
item to signify a change in opinion (but / however / although).