Add support for training deep models with different likelihoods
In order to experiment with different ways of leveraging the
redundant annotations for stronger supervision, extend the training
scripts to support the use of different likelihoods: cross-entropy
on the hard labels, cross-entropy on soft / averaged labels,
cross-entropy on all the annotations, and a dirichlet-multinomial
likelihood.

Since the models are trained with different likelihoods, we also
need different logic for computing class probabilities from these
models, so the prediction scripts have to be updated.