Add functions to estimate beta-binomials and dirichlet-multinomials
To compute the oracle cross-entropy score, we'll need to be able
to fit beta-binomial and dirichlet-multinomial models to the data.
Fitting these models via MLE allows us to make an empirical
Bayesian estimate for the minimum possible cross-entropy score
i.e. the cross-entropy obtained by always predicting with the
true underlying distribution.