Add a script to compute the oracle performance
Add a script to estimate a lower bound on how well any model can
do. We compute this lower bound using an empirical Bayesian
methodology, first learning a prior for the label distribution of
an instance, and then sampling "true" label distributions from the
posterior and averaging over them to estimate the oracle performance
(i.e. how well a model would do if it learned the conditional label
distribution perfectly).