Update pytorch_pretrained_bert to transformers
The transformers library is the successsor to
pytorch_pretrained_bert. It includes implementations of newer
models like RoBERTa. Updating to transformers requires switching
from the BertAdam and warmup_linear APIs to the AdamW and
WarmupLinearSchedule APIs. These APIs allow us to drop some of the
special casing logic for mixed precision if we also upgrade to the
newer amp API for apex. So, this commit implements all these changes.