Suppress all warnings from pytest output
This code is primarily for reproducibility and archival purposes.
The tests are only intended to give people some confidence that
their installation of the code works as it originally did in the
Scruples research project. Currently, the deprecation and related
warnings create so much noise that the tests cease to be useful.